+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.2] - 2024-01-31
+
+- Added test to ensure that bUseUObjectArrayCache is set to false regardless of game store (on extension activation).
+- Modified UE4SS installer to disable the ObjectArrayCache on scripting system installation.
+
 ## [0.1.1] - 2024-01-26
 
 - Added support for Blueprint mods which are deployed into ../Paks/LogicMods/